refector: clean up user service

+++ src/app/modules/user/user.service.ts
@@ -3,9 +3,8 @@ import { Prisma } from "../../../generated/client";
 import { prisma } from "../../../lib/prisma"
 
 const getAllUsers = async () => {
-    const users = await prisma.user.findMany();
-    const totalUsers = await prisma.user.count();
-    return { total: totalUsers, data: users }
+    return await prisma.user.findMany();
+
 }
 
 const getSingleUser = async (id: string) => {
@@ -16,6 +15,7 @@ const updateUser = async (id: string, updateData: Prisma.UserUpdateInput) => {
     const isExistUser = await prisma.user.findFirst({ where: { id } })
     if (!isExistUser) { throw new Error("user is not exists") }
 
+
     return await prisma.user.update({
         where: { id },
         data: updateData,
@@ -25,6 +25,17 @@ const updateUser = async (id: string, updateData: Prisma.UserUpdateInput) => {
 const deleteUser = async (id: string) => {
     const isExistUser = await prisma.user.findUnique({ where: { id } })
     if (!isExistUser) { throw new Error("user is not exist in DB") }
+
+    const orderCount = await prisma.order.count({
+        where: { userId: id }
+    })
+
+    if (orderCount) {
+        throw new Error(
+            "This user cannot be deleted because they have existion orders. Please Inactive the user instead"
+        )
+    }
+
     return await prisma.user.delete({ where: { id } })
 }
